Some minor changes

+++ src/utils.cpp
@@ -6,12 +6,12 @@
 
 using namespace std;
 
-int bitVecToInt(vector<bit> vec) {
+unsigned int bitVecToInt(vector<bit> vec) {
     /* Converts a vector of bits to an integer */
-    int result = 0, size = vec.size();
+    unsigned int result = 0, size = vec.size();
     reverse(vec.begin(), vec.end());
 
-    for (int i = 0; i < size; ++i)
+    for (size_t i = 0; i < size; ++i)
         result += vec[i] * pow(2, i);
 
     return result;
